Add agent note about EchoLLM relocation
