+++ src/entity/resources/interfaces/__init__.py
@@ -1,6 +1,7 @@
 from .database import DatabaseResource
 from .duckdb_resource import DuckDBResource
 from .duckdb_vector_store import DuckDBVectorStore
+from .echo_llm import EchoLLMResource
 from .llm import LLMResource
 from .storage import StorageResource
 from .vector_store import VectorStoreResource
@@ -10,6 +11,7 @@ __all__ = [
     "VectorStoreResource",
     "DuckDBVectorStore",
     "DuckDBResource",
+    "EchoLLMResource",
     "LLMResource",
     "StorageResource",
 ]
