+++ src/entity/resources/interfaces/echo_llm.py
