+++ src/plugins/builtin/resources/__init__.py
@@ -1,11 +1,9 @@
 """Resource implementations shipped with the framework."""
 
-from .echo_llm import EchoLLMResource
 from .ollama_llm import OllamaLLMResource
 from .pg_vector_store import PgVectorStore
 
 __all__ = [
-    "EchoLLMResource",
     "OllamaLLMResource",
     "PgVectorStore",
 ]
